third_party/blink/web_tests/printing/page-rule-selection.html

<!DOCTYPE html>
<html id="html_element">
<head id="head_element">
<script src="../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
    description("This tests page style selection.");

    function appendStyle(styleString)
    {
        var styleElement = document.createElement("style");
        styleElement.innerHTML = styleString;
        document.getElementById("head_element").appendChild(styleElement);
    }

    if (window.testRunner) {
        testRunner.dumpAsText();

        debug("Check font update and line height override.");
        appendStyle("@page { line-height:1em; }");
        appendStyle("@page { font-family:Futura; }");
        appendStyle("@page { font:15px/6em Arial; }");
        appendStyle("@page { zoom:200%; }");
        shouldBeEqualToString("internals.pageProperty('line-height', 0)", "180");
        shouldBeEqualToString("internals.pageProperty('font-family', 0)", "Arial");
        shouldBeEqualToString("internals.pageProperty('font-size', 0)", "30");
        appendStyle("@page { zoom:100%; }");

        debug("Check @media rules.");
        appendStyle("@page { size:100px 200px; }");
        shouldBeEqualToString("internals.pageProperty('size', 0)", "100 200");
        appendStyle("@media print { @page { size:150px 250px; } }");
        shouldBeEqualToString("internals.pageProperty('size', 0)", "150 250");
        appendStyle("@media screen { @page { size:300px 500px; } }");
        shouldBeEqualToString("internals.pageProperty('size', 0)", "150 250");

    } else {
        testFailed("This test can be run only with window.testRunner");
    }
</script>
</body>
</html>
